Add files via upload
A tuple refers to a data structure that consists of an ordered set of elements. Tuples can be used to store multiple values that are related to each other and maintain their order, making them useful for representing structured data. the elements of a tuple can be of different data types, such as strings, integers, or other data structures.